backend/requirements.txt
fastapi
uvicorn
jira
google-generativeai
python-dotenv
pydantic
python-multipart
requests
